requirements.txt
requests == 2.24.0
pyaml == 20.4.0